mbspex: driver - added internal sfp and slave broadcast for writing values gosipcmd accepts -1 for sfp and slave to indicate such broadcast configuration files for mbs poland linux examples s2stairs with single addressing and broadcast mode
git-svn-id: https://subversion.gsi.de/dabc/drivers@2303 bcbf6573-9a26-0410-9ebc-ce4ab7aade96

+++ mbspex/driver/pex_user.h
@@ -53,8 +53,8 @@ struct pex_dma_io {
 };
 
 struct pex_bus_io {
-	unsigned long sfp;		/* sfp link id 0..3 (optional)*/
-	unsigned long slave;	/* slave device id at the sfp (optional)*/
+	int sfp;		/* sfp link id 0..3 (-1 for broadcast to all configured sfps)*/
+	long slave;	    /* slave device id at the sfp (-1 ato broadcast to all slaves)*/
 	unsigned long address;	/* address on the "field bus" connected to the optical links*/
 	unsigned long value;	/* value for read/write at bus address. Contains result status after write*/
 };
